services/pyproject.toml
[project]
name = "services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.2.0",
    "asyncmy>=0.2.10",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "mysqlclient>=2.2.7",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
]
